requirements.txt
Django==5.0.6
djangorestframework==3.15.2
mysqlclient==2.2.7

# Ferramentas úteis
python-dotenv==1.0.1
django-cors-headers==4.3.1

djangorestframework-simplejwt
django-cors-headers